main/authentication/requirements.txt
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
dialogflow==1.0.0
google-api-core==1.17.0
google-api-python-client==1.8.4
google-auth==1.15.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
grpcio==1.29.0
httplib2==0.18.1
idna==2.9
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
requests==2.23.0
rsa==4.0
six==1.15.0
uritemplate==3.0.1
urllib3==1.25.9
uuid==1.30
